+++ eth-types/src/lib.rs
@@ -61,6 +61,30 @@ use crate::evm_types::Stack;
 #[cfg(feature = "enable-storage")]
 use crate::evm_types::Storage;
 
+/// Used for FFI with Golang. Bytes in golang will be serialized as base64 by default.
+pub mod base64 {
+    use base64::{decode, encode};
+    use serde::{Deserialize, Deserializer, Serialize, Serializer};
+
+    /// serialize bytes as base64
+    pub fn serialize<S>(data: &[u8], s: S) -> Result<S::Ok, S::Error>
+    where
+        S: Serializer,
+    {
+        String::serialize(&encode(data), s)
+    }
+
+    /// deserialize base64 to bytes
+    pub fn deserialize<'de, D>(d: D) -> Result<Vec<u8>, D::Error>
+    where
+        D: Deserializer<'de>,
+    {
+        let s = String::deserialize(d)?;
+        decode(s.as_bytes()).map_err(serde::de::Error::custom)
+    }
+}
+
+
 /// Trait used to reduce verbosity with the declaration of the [`Field`]
 /// trait and its repr.
 pub trait Field:
