+++ prover/src/types.rs
@@ -9,23 +9,4 @@ pub type WitnessBlock = Block<Fr>;
 pub struct BlockTraceJsonRpcResult {
     pub result: BlockTrace,
 }
-
-pub mod base64 {
-    use base64::{decode, encode};
-    use serde::{Deserialize, Deserializer, Serialize, Serializer};
-
-    pub fn serialize<S>(data: &[u8], s: S) -> Result<S::Ok, S::Error>
-    where
-        S: Serializer,
-    {
-        String::serialize(&encode(data), s)
-    }
-
-    pub fn deserialize<'de, D>(d: D) -> Result<Vec<u8>, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        let s = String::deserialize(d)?;
-        decode(s.as_bytes()).map_err(serde::de::Error::custom)
-    }
-}
+pub use eth_types::base64;
